more

+++ db/migrate/20250919101114_create_schools_manual_readings.rb
@@ -3,8 +3,8 @@ class CreateSchoolsManualReadings < ActiveRecord::Migration[7.2]
     create_table :schools_manual_readings do |t|
       t.references :school, null: false, foreign_key: { on_delete: :cascade }
       t.date :month, null: false
-      t.integer :electricity
-      t.integer :gas
+      t.float :electricity
+      t.float :gas
 
       t.timestamps
